front end delete friend, untested

+++ server/routes/relationships.ts
@@ -6,21 +6,27 @@ import * as db from '../db/relationships.ts'
 
 const router = Router()
 
-router.delete('/:currentUserId/:friendId', async (req, res) => {
-  // if (!req.auth?.sub) {
-  //   res.sendStatus(StatusCodes.UNAUTHORIZED)
-  //   return
-  // }
-  try {
-    const { currentUserId, friendId } = req.params
-    const isDeleted = await db.deleteRelationship(currentUserId, friendId)
-    isDeleted
-      ? res.sendStatus(StatusCodes.NO_CONTENT)
-      : res.sendStatus(StatusCodes.NOT_FOUND)
-  } catch (err) {
-    console.error(err instanceof Error ? err.message : 'Error deleting message')
-    res.sendStatus(StatusCodes.INTERNAL_SERVER_ERROR)
-  }
-})
+router.delete(
+  '/:currentUserId/:friendId',
+  checkJwt,
+  async (req: JwtRequest, res) => {
+    if (!req.auth?.sub) {
+      res.sendStatus(StatusCodes.UNAUTHORIZED)
+      return
+    }
+    try {
+      const { currentUserId, friendId } = req.params
+      const isDeleted = await db.deleteRelationship(currentUserId, friendId)
+      isDeleted
+        ? res.sendStatus(StatusCodes.NO_CONTENT)
+        : res.sendStatus(StatusCodes.NOT_FOUND)
+    } catch (err) {
+      console.error(
+        err instanceof Error ? err.message : 'Error deleting message',
+      )
+      res.sendStatus(StatusCodes.INTERNAL_SERVER_ERROR)
+    }
+  },
+)
 
 export default router
